Potentially chain multiple handlers
There is a list of handlers, each with an analyse() function and a
handle() function. If analyse() returns a truthy value, then this
handler is used (its handle() called, with the result of analyse() to
facilitate data transfer between both phases), otherwise the next
handler is tried. Right now only a single handler is registered, a
fallback which always matches and proxies to the Wikidata Query Service.