Set up nodemon
nodemon [1] watches the file system for changes to the source files and
automatically restarts the program. This is very useful for local
development – run `npm start` and make code changes as you like, and
just forget about the server running somewhere in the background. (If
the server doesn’t respond at all – in the browser, the requests hang –
then it probably crashed with an error. nodemon will restart it on the
next code change, or you can enter "rs" and then Enter at the nodemon
prompt to restart it manually.)

[1]: https://github.com/remy/nodemon