Implement and register OptimizingHandler
This handler proxies to WDQS after optimizing the query at the SPARQL
level. It currently implements a single optimization: a query like

    SELECT ?item ?dod WHERE {
      ?item wdt:P570 ?dod.
      FILTER(YEAR(?dod) = 2020)
    }
    LIMIT 100

is rewritten to the equivalent, but more efficient [1]

    SELECT * WHERE {
      ?item wdt:P570 ?dod.
      FILTER((?dod >= "2020-00-00"^^xsd:dateTime) && (?dod < "2021-00-00"^^xsd:dateTime))
    }
    LIMIT 100

In this case, that seems to speed up the query by roughly one order of
magnitude (from 45-50 s to 4-5 s).

[1]: https://twitter.com/WikidataFacts/status/752187548219412480