Start adding LabelServiceHandler
So far, only the analysis is done: detect cases where we can safely
split the label service off from the rest of the query, and run it
afterwards as a separate step, using either prop=entityterms (if no
language fallbacks are requested) or action=wbgetentities (if fallbacks
to English are requested). For now, just dump the analysis results to
stdout, the rest of the implementation will come later.