Add rudimentary stats module
Count how often each request type is used, and make the statistics
available on the /queripulator-stats route, as well as dumping them to
stdout on graceful exit (because the Cloud VPS service restarts often,
so without saving the stats to the journal there won’t be much to
report). Note that this doesn’t distinguish between WdqsRequest objects
from SpecificItemPropertyPairQueryHandler, OptimizingHandler, or the
default fallback handler yet.