Extract WdqsPlusApiRequest base class
This code was completely duplicated between EntitytermsRequest and
WbgetentitiesRequest and can be extracted quite easily. As an added
bonus, now that the Transform is created in the constructor (when
calling the parent constructor), we don’t even need to store the
variableMappings and language in properties now (since they’re only
really needed by the Transform, not the Request).