Add TruthyHandler
This detects when a query uses only truthy triples. The idea is that
such a query could be sent to a different query service (cluster), which
holds a smaller amount of data and may be more efficient.

The defaultQueryHandlerChain is reordered overall, because
SubjectByPropertyValueHandler needs to come before TruthyHandler,
otherwise it’ll never match (any subject-by-property query is also
truthy). This probably makes sense in general, too – I’m not sure why
SubjectByPropertyValueHandler was so far down the chain before.