Work around another ldf-client bug
ldf-client also produces an empty head section when the query uses
SELECT * (even though it shouldn’t be too difficult for ldf-client to
collect all the variables used in the query and put all of them in the
head). To work around this, don’t use ldf-client for such queries
either.