Don’t use axios for WdqsRequest
The main motivation for this is to forward all the request headers to
WDQS, because they’re probably useful. This includes Accept-Encoding,
which lets the query server respond with compressed data, which can be
useful; however, that is also why we have to stop using axios: axios
“transparently” decompresses the response, but does not remove or adjust
the Content-Length header, so the browser will see a truncated response.

This effectively reverts bf0e2cf75e.