Allow label service in truthy handler
I had assumed the truthy dumps did not contain labels, descriptions, or
aliases, and therefore the truthy service would not be able to support
the label service. But in fact it looks like those terms are all
included, so as far as I understand there’s little reason to exclude the
label service.